Refactor quickcheck tests to use common function
